Migarate `xpallet_gateway_bitcoin` to pallet v2 (#532)
* Migarate `xpallet_gateway_bitcoin` to pallet v2

* Reorg crate structure

* Add missing dir

* Add space line

* Make clippy happy

* Explicate number type
