+++ Cargo.toml
@@ -9,7 +9,7 @@ description = "Fully Decentralized Interchain Crypto Asset Management on Polkado
 cli = { package = "chainx-cli", path = "cli" }
 
 [target.'cfg(target_arch="x86_64")'.dependencies]
-cli = { package = "chainx-cli", path = "cli", features = [ "wasmtime" ] }
+cli = { package = "chainx-cli", path = "cli", features = ["wasmtime"] }
 
 [patch.crates-io]
 sc-authority-discovery = { git = "https://github.com/paritytech/substrate", rev = "e6ac7e72c2251b8a23454c8cf0e3a2ad2c1ed566" }
@@ -101,53 +101,52 @@ pallet-utility = { git = "https://github.com/paritytech/substrate", rev = "e6ac7
 
 [workspace]
 members = [
-    "cli",
-    "executor",
-    "primitives",
-    "primitives/assets-registrar",
-    "primitives/gateway/bitcoin",
-    "primitives/gateway/common",
-    "primitives/genesis-builder",
-    "primitives/io",
-    "primitives/logging",
-    "primitives/mining/common",
-    "primitives/mining/staking",
-    "primitives/protocol",
-    "primitives/rpc",
-    "primitives/runtime",
-    "rpc",
-    "runtime/chainx",
-    "runtime/common",
-    "runtime/dev",
-    "runtime/malan",
-    "service",
-
-    "xpallets/assets",
-    "xpallets/assets/rpc",
-    "xpallets/assets/rpc/runtime-api",
-    "xpallets/assets-registrar",
-    "xpallets/dex/spot",
-    "xpallets/dex/spot/rpc",
-    "xpallets/dex/spot/rpc/runtime-api",
-    "xpallets/gateway/bitcoin",
-    "xpallets/gateway/common",
-    "xpallets/gateway/common/rpc",
-    "xpallets/gateway/common/rpc/runtime-api",
-    "xpallets/gateway/records",
-    "xpallets/gateway/records/rpc",
-    "xpallets/gateway/records/rpc/runtime-api",
-    "xpallets/genesis-builder",
-    "xpallets/mining/asset",
-    "xpallets/mining/asset/rpc",
-    "xpallets/mining/asset/rpc/runtime-api",
-    "xpallets/mining/staking",
-    "xpallets/mining/staking/rpc",
-    "xpallets/mining/staking/rpc/runtime-api",
-    "xpallets/support",
-    "xpallets/system",
-    "xpallets/transaction-fee",
-    "xpallets/transaction-fee/rpc",
-    "xpallets/transaction-fee/rpc/runtime-api",
+  "cli",
+  "executor",
+  "primitives",
+  "primitives/assets-registrar",
+  "primitives/gateway/bitcoin",
+  "primitives/gateway/common",
+  "primitives/genesis-builder",
+  "primitives/io",
+  "primitives/logging",
+  "primitives/mining/common",
+  "primitives/mining/staking",
+  "primitives/protocol",
+  "primitives/rpc",
+  "primitives/runtime",
+  "rpc",
+  "runtime/chainx",
+  "runtime/common",
+  "runtime/dev",
+  "runtime/malan",
+  "service",
+  "xpallets/assets",
+  "xpallets/assets/rpc",
+  "xpallets/assets/rpc/runtime-api",
+  "xpallets/assets-registrar",
+  "xpallets/dex/spot",
+  "xpallets/dex/spot/rpc",
+  "xpallets/dex/spot/rpc/runtime-api",
+  "xpallets/gateway/bitcoin/v1",
+  "xpallets/gateway/common",
+  "xpallets/gateway/common/rpc",
+  "xpallets/gateway/common/rpc/runtime-api",
+  "xpallets/gateway/records",
+  "xpallets/gateway/records/rpc",
+  "xpallets/gateway/records/rpc/runtime-api",
+  "xpallets/genesis-builder",
+  "xpallets/mining/asset",
+  "xpallets/mining/asset/rpc",
+  "xpallets/mining/asset/rpc/runtime-api",
+  "xpallets/mining/staking",
+  "xpallets/mining/staking/rpc",
+  "xpallets/mining/staking/rpc/runtime-api",
+  "xpallets/support",
+  "xpallets/system",
+  "xpallets/transaction-fee",
+  "xpallets/transaction-fee/rpc",
+  "xpallets/transaction-fee/rpc/runtime-api",
 ]
 
 [profile.release]
