+++ xpallets/gateway/bitcoin/v1/Cargo.toml
@@ -26,17 +26,17 @@ pallet-timestamp = { version = "3.0.0", default-features = false }
 orml-utilities = { version = "0.4.0", default-features = false }
 
 # ChainX primitives
-chainx-primitives = { path = "../../../primitives", default-features = false }
-xp-gateway-bitcoin = { path = "../../../primitives/gateway/bitcoin", default-features = false }
-xp-gateway-common = { path = "../../../primitives/gateway/common", default-features = false }
-xp-logging = { path = "../../../primitives/logging", default-features = false }
-xp-protocol = { path = "../../../primitives/protocol", default-features = false }
+chainx-primitives = { path = "../../../../primitives", default-features = false }
+xp-gateway-bitcoin = { path = "../../../../primitives/gateway/bitcoin", default-features = false }
+xp-gateway-common = { path = "../../../../primitives/gateway/common", default-features = false }
+xp-logging = { path = "../../../../primitives/logging", default-features = false }
+xp-protocol = { path = "../../../../primitives/protocol", default-features = false }
 
 # ChainX pallets
-xpallet-assets = { path = "../../assets", default-features = false }
-xpallet-gateway-common = { path = "../common", default-features = false }
-xpallet-gateway-records = { path = "../records", default-features = false }
-xpallet-support = { path = "../../support", default-features = false }
+xpallet-assets = { path = "../../../assets", default-features = false }
+xpallet-gateway-common = { path = "../../common", default-features = false }
+xpallet-gateway-records = { path = "../../records", default-features = false }
+xpallet-support = { path = "../../../support", default-features = false }
 
 # light-bitcoin
 light-bitcoin = { git = "https://github.com/chainx-org/light-bitcoin", branch = "master", default-features = false }
@@ -50,40 +50,40 @@ sp-keyring = "3.0.0"
 frame-benchmarking = "3.0.0"
 pallet-balances = "3.0.0"
 pallet-multisig = "3.0.0"
-xp-assets-registrar = { path = "../../../primitives/assets-registrar" }
-xpallet-assets-registrar = { path = "../../assets-registrar" }
+xp-assets-registrar = { path = "../../../../primitives/assets-registrar" }
+xpallet-assets-registrar = { path = "../../../assets-registrar" }
 
 [features]
 default = ["std"]
 std = [
-    "bs58/std",
-    "codec/std",
-    "hex/std",
-    "serde",
-    # Substrate primitives
-    "sp-core/std",
-    "sp-io/std",
-    "sp-runtime/std",
-    "sp-std/std",
-    # Substrate pallets
-    "frame-support/std",
-    "frame-system/std",
-    "pallet-timestamp/std",
-    # Orml
-    "orml-utilities/std",
-    # ChainX primitives
-    "chainx-primitives/std",
-    "xp-gateway-bitcoin/std",
-    "xp-gateway-common/std",
-    "xp-logging/std",
-    "xp-protocol/std",
-    # ChainX pallets
-    "xpallet-assets/std",
-    "xpallet-gateway-common/std",
-    "xpallet-gateway-records/std",
-    "xpallet-support/std",
-    # light-bitcoin
-    "light-bitcoin/std",
+  "bs58/std",
+  "codec/std",
+  "hex/std",
+  "serde",
+  # Substrate primitives
+  "sp-core/std",
+  "sp-io/std",
+  "sp-runtime/std",
+  "sp-std/std",
+  # Substrate pallets
+  "frame-support/std",
+  "frame-system/std",
+  "pallet-timestamp/std",
+  # Orml
+  "orml-utilities/std",
+  # ChainX primitives
+  "chainx-primitives/std",
+  "xp-gateway-bitcoin/std",
+  "xp-gateway-common/std",
+  "xp-logging/std",
+  "xp-protocol/std",
+  # ChainX pallets
+  "xpallet-assets/std",
+  "xpallet-gateway-common/std",
+  "xpallet-gateway-records/std",
+  "xpallet-support/std",
+  # light-bitcoin
+  "light-bitcoin/std",
 ]
 runtime-benchmarks = ["frame-benchmarking"]
 ss58check = ["xp-gateway-common/ss58check"]
