+++ xpallets/gateway/bitcoin/v1/src/header/header_proof.rs
