+++ xpallets/gateway/bitcoin/v1/src/tx/secp256k1_verifier/recover_verifier/scalar.rs
