+++ xpallets/gateway/bitcoin/v1/src/tx/validator.rs
