+++ xpallets/gateway/bitcoin/v1/src/types.rs
