+++ xpallets/gateway/records/src/benchmarking.rs
@@ -21,7 +21,7 @@ fn deposit<T: Config>(who: T::AccountId, amount: BalanceOf<T>) {
 
 fn deposit_and_withdraw<T: Config>(who: T::AccountId, amount: BalanceOf<T>) {
     deposit::<T>(who.clone(), amount);
-    let withdrawal = amount - 500.into();
+    let withdrawal = amount - 500u32.into();
     let addr = b"3LFSUKkP26hun42J1Dy6RATsbgmBJb27NF".to_vec();
     let memo = b"memo".to_vec().into();
     let receiver_lookup: <T::Lookup as StaticLookup>::Source = T::Lookup::unlookup(who);
@@ -53,9 +53,9 @@ benchmarks! {
     root_withdraw {
         let receiver: T::AccountId = whitelisted_caller();
         let receiver_lookup: <T::Lookup as StaticLookup>::Source = T::Lookup::unlookup(receiver.clone());
-        let amount: BalanceOf<T> = 1000.into();
+        let amount: BalanceOf<T> = 1000u32.into();
         deposit::<T>(receiver, amount);
-        let withdrawal = amount - 500.into();
+        let withdrawal = amount - 500u32.into();
         let addr = b"3LFSUKkP26hun42J1Dy6RATsbgmBJb27NF".to_vec();
         let memo = b"memo".to_vec().into();
     }: _(RawOrigin::Root, receiver_lookup, ASSET_ID, withdrawal, addr, memo)
